RDM-10414 - Refactoring
