--- src/main/java/uk/gov/hmcts/ccd/ApplicationParams.java
@@ -1,7 +1,0 @@
-package uk.gov.hmcts.ccd;
-
-import org.springframework.stereotype.Component;
-
-@Component
-public class ApplicationParams {
-}
