+++ src/main/java/uk/gov/hmcts/ccd/config/MessagePublisherParams.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.ccd.service.messaging;
+package uk.gov.hmcts.ccd.config;
 
 import lombok.Data;
 import org.springframework.boot.context.properties.ConfigurationProperties;
