+++ src/main/java/uk/gov/hmcts/ccd/config/PublishMessageTask.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.ccd.service.messaging;
+package uk.gov.hmcts.ccd.config;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
