+++ src/main/java/uk/gov/hmcts/ccd/service/MessagePublisherRunnable.java
@@ -1,10 +1,11 @@
-package uk.gov.hmcts.ccd.service.messaging;
+package uk.gov.hmcts.ccd.service;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 import org.springframework.jms.core.JmsTemplate;
+import uk.gov.hmcts.ccd.config.PublishMessageTask;
 import uk.gov.hmcts.ccd.data.MessageQueueCandidateEntity;
 import uk.gov.hmcts.ccd.data.MessageQueueCandidateRepository;
 
@@ -19,6 +20,7 @@ public class MessagePublisherRunnable implements Runnable {
     private MessageQueueCandidateRepository messageQueueCandidateRepository;
     private JmsTemplate jmsTemplate;
     private PublishMessageTask publishMessageTask;
+    private String logPrefix;
 
     public MessagePublisherRunnable(MessageQueueCandidateRepository messageQueueCandidateRepository,
                                     JmsTemplate jmsTemplate,
@@ -26,12 +28,15 @@ public class MessagePublisherRunnable implements Runnable {
         this.messageQueueCandidateRepository = messageQueueCandidateRepository;
         this.jmsTemplate = jmsTemplate;
         this.publishMessageTask = publishMessageTask;
+        this.logPrefix = String.format("[Message Type - %s]", publishMessageTask.getMessageType());
     }
 
     @Override
     public void run() {
+        log.debug(String.format("%s Starting publish message task", logPrefix));
         processUnpublishedMessages(PageRequest.of(0, publishMessageTask.getBatchSize()), newArrayList());
         deletePublishedMessages();
+        log.debug(String.format("%s Completed publish message task", logPrefix));
     }
 
     private void processUnpublishedMessages(Pageable pageable,
@@ -44,8 +49,8 @@ public class MessagePublisherRunnable implements Runnable {
                 .findUnpublishedMessages(publishMessageTask.getMessageType(), pageable);
             publishMessages(unpublishedMessagesPaginated, processedEntities);
         } catch (Exception e) {
-            log.error(String.format("Error encountered during processing of unpublished messages for "
-                + "message type '%s'", publishMessageTask.getMessageType()), e);
+            log.error(String.format("%s Error encountered during processing of "
+                + "unpublished messages", logPrefix), e);
             hasError = true;
         }
 
@@ -53,8 +58,8 @@ public class MessagePublisherRunnable implements Runnable {
             processUnpublishedMessages(unpublishedMessagesPaginated.nextPageable(), processedEntities);
         } else if (!processedEntities.isEmpty()) {
             messageQueueCandidateRepository.saveAll(processedEntities);
-            log.info(String.format("[Message Type - %s] Published %s messages to destination '%s'",
-                publishMessageTask.getMessageType(), processedEntities.size(), publishMessageTask.getDestination()));
+            log.info(String.format("%s Published %s messages to destination '%s'",
+                logPrefix, processedEntities.size(), publishMessageTask.getDestination()));
         }
     }
 
@@ -71,7 +76,7 @@ public class MessagePublisherRunnable implements Runnable {
         LocalDateTime retentionDate = LocalDateTime.now().minusDays(publishMessageTask.getPublishedRetentionDays());
         int result = messageQueueCandidateRepository
             .deletePublishedMessages(retentionDate, publishMessageTask.getMessageType());
-        log.debug(String.format("[Message Type - '%s'] Deleted %s records with publish date before %s",
-            publishMessageTask.getMessageType(), result, retentionDate.toString()));
+        log.debug(String.format("%s Deleted %s records with publish date before %s",
+            logPrefix, result, retentionDate.toString()));
     }
 }
