+++ src/main/resources/application.yaml
@@ -30,8 +30,8 @@ spring:
     properties:
       charSet: UTF-8
     hikari:
-      minimumIdle: ${MESSAGE_PUBLISHER_DB_MIN_IDLE:4}
-      maximumPoolSize: ${MESSAGE_PUBLISHER_DB_MAX_POOL_SIZE:12}
+      minimumIdle: ${MESSAGE_PUBLISHER_DB_MIN_IDLE:2}
+      maximumPoolSize: ${MESSAGE_PUBLISHER_DB_MAX_POOL_SIZE:8}
       idleTimeout: ${MESSAGE_PUBLISHER_DB_IDLE_TIMEOUT:300000}
       maxLifetime: ${MESSAGE_PUBLISHER_DB_MAX_LIFTIME:7200000}
       connectionTimeout: ${MESSAGE_PUBLISHER_DB_CONNECTION_TIMEOUT:40000}
